Day3 - Part2 fixed integer overflow bug

+++ src/main/kotlin/challenges/day3/day3.kt
@@ -11,11 +11,11 @@ fun execute(){
     val numberOfTrees = part1(1, 3, grid)
     println("Part1: $numberOfTrees")
 
-    val r1d1 = part1(1, 1, grid)
-    val r3d1 = part1(1, 3, grid)
-    val r5d1 = part1(1, 5, grid)
-    val r7d1 = part1(1, 7, grid)
-    val r1d2 = part1(2, 1, grid)
+    val r1d1 = part1(1, 1, grid).toLong()
+    val r3d1 = part1(1, 3, grid).toLong()
+    val r5d1 = part1(1, 5, grid).toLong()
+    val r7d1 = part1(1, 7, grid).toLong()
+    val r1d2 = part1(2, 1, grid).toLong()
     println("Part2: ${r1d1*r3d1*r5d1*r7d1*r1d2}")
 }
